+++ src/TransitionKind.ts
@@ -1,0 +1,13 @@
+/*
+ * Finite state machine library
+ * Copyright (c) 2014-5 Steelbreeze Limited
+ * Licensed under the MIT and GPL v3 licences
+ * http://www.steelbreeze.net/state.cs
+ */
+module StateJS {	
+	export enum TransitionKind {		
+		Internal,
+		Local,
+		External
+	}
+}